Refactored the Player class
Simplified the physics.
Player now uses a state machine.
Player Y coordinate 0 is now at the bottom of the level with axis going up instead of down.
Player can notify something when it moves (We use this to update the camera position).
CollideRect  and Center can be set externally (and Center is in pixels instead of decimals).
Player plays a different animation based on direction.